requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Jinja2==3.1.6
passlib[bcrypt]==1.7.4
email_validator==2.2.0
python-jose[cryptography]==3.3.0
requests==2.32.3
pyjwt==2.8.1
cryptography==41.0.7


